- creat metoda add amount

+++ src/main/java/com/sda/practical/project/model/AccountsModel.java
@@ -42,7 +42,7 @@ public class AccountsModel {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
